refactor(ci): deepen GitHub ownership
