+++ scripts/dependencies/bump.ts
@@ -1,11 +1,11 @@
 import { NodeRuntime, NodeServices } from "@effect/platform-node";
 import { Config, Effect, Layer, Result, Schema } from "effect";
 import { FetchHttpClient } from "effect/unstable/http";
+import { inspectGithubActionPolicy } from "../github/policy.ts";
 import {
   fetchLatestGithubActionTag,
   githubActionReleaseReviews,
-  inspectGithubActionPolicy,
-} from "../github-action-policy.ts";
+} from "../github/release.ts";
 import { writeError, writeOutput } from "../output.ts";
 import { runPnpm } from "./command.ts";
 import { REGISTRY_REVIEWS } from "./policy.ts";
@@ -195,7 +195,7 @@ export const bumpDependencies = Effect.fn("RepositoryPolicy.bumpDependencies")(
     }
 
     yield* writeOutputMessage(
-      "Routine dependencies and every reviewed hold are current under the repository's 24-hour release-maturity policy and exact reviewed exception allowlist.\n"
+      "Routine dependencies and every reviewed hold match the exact repository policy and reviewed exception allowlist.\n"
     );
     return 0;
   }
