+++ scripts/github/release.test.ts
@@ -1,0 +1,79 @@
+import { describe, expect, it } from "@effect/vitest";
+import { Effect, Layer, Result } from "effect";
+import {
+  HttpClient,
+  type HttpClientRequest,
+  HttpClientResponse,
+} from "effect/unstable/http";
+import {
+  fetchLatestGithubActionTag,
+  githubActionReleaseReviews,
+} from "./release.ts";
+
+function makeHttpClient(
+  makeResponse: (request: HttpClientRequest.HttpClientRequest) => Response
+) {
+  return Layer.succeed(
+    HttpClient.HttpClient,
+    HttpClient.make((request) =>
+      Effect.sync(() =>
+        HttpClientResponse.fromWeb(request, makeResponse(request))
+      )
+    )
+  );
+}
+
+describe("GitHub Action releases", () => {
+  it.effect("deduplicates reviews by upstream repository", () =>
+    Effect.gen(function* () {
+      const reviews = yield* githubActionReleaseReviews();
+      const repositories = reviews.map(({ repository }) => repository);
+
+      expect(new Set(repositories).size).toBe(repositories.length);
+      expect(repositories).toContain("actions/cache");
+    })
+  );
+
+  it.effect("reads release metadata through the Effect HTTP client", () =>
+    Effect.gen(function* () {
+      let observedRequest: HttpClientRequest.HttpClientRequest | undefined;
+      const releaseTag = yield* fetchLatestGithubActionTag({
+        repository: "actions/checkout",
+      }).pipe(
+        Effect.provide(
+          makeHttpClient((request) => {
+            observedRequest = request;
+            return Response.json({ tag_name: "v7.0.1" });
+          })
+        )
+      );
+
+      expect(releaseTag).toBe("v7.0.1");
+      expect(observedRequest?.url).toBe(
+        "https://api.github.com/repos/actions/checkout/releases/latest"
+      );
+    })
+  );
+
+  it.effect("rejects unavailable GitHub release metadata", () =>
+    Effect.gen(function* () {
+      const result = yield* fetchLatestGithubActionTag({
+        repository: "actions/checkout",
+      }).pipe(
+        Effect.provide(
+          makeHttpClient(() => new Response(null, { status: 403 }))
+        ),
+        Effect.result
+      );
+
+      expect(Result.isFailure(result)).toBe(true);
+      if (Result.isSuccess(result)) {
+        return;
+      }
+      expect(result.failure).toMatchObject({
+        _tag: "GithubActionReleaseError",
+        message: "Unable to read the latest actions/checkout release.",
+      });
+    })
+  );
+});
